fix iso1:PR Puerto Rico
